Centralización de Skeletons (components/ui/skeletons/) incluye: BookCardSkeleton.tsxTarjeta de libro (grid), PostCardSkeleton.tsxTarjeta de publicación de comunidad. UserCardSkeleton.tsxTarjeta de usuario (búsqueda), CommunityCardSkeleton.tsxTarjeta de comunidad (búsqueda), RecommendationSkeleton.tsxItem de recomendación lateral, BookDetailSkeleton.tsxPantalla completa de detalle de libro, ChatSkeleton.tsxBurbujas de chat
