+++ app/(dashboard)/communities/[id]/CommunityDetailClient.tsx
@@ -20,6 +20,7 @@ import { X, Upload, Users, MessageSquare, Search, Settings, Trash2, Camera, LogO
 import ImageCropper, { type AspectRatioOption } from "@/components/ui/ImageCropper";
 import { fileToDataUrl, blobToFile } from "@/lib/imageUtils";
 import { BOOK_GENRES } from "@/lib/constants/genres";
+import { PostCardSkeleton } from "@/components/ui/skeletons";
 
 const POST_ASPECT_RATIOS: AspectRatioOption[] = [
     { label: "1:1", value: 1, icon: "square" },
@@ -29,40 +30,7 @@ const POST_ASPECT_RATIOS: AspectRatioOption[] = [
     { label: "2:3", value: 2 / 3, icon: "portrait" },
 ];
 
-// ─── Skeleton Components ─────────────────────────────────────────────────────
 
-function PostCardSkeleton() {
-    return (
-        <div className="bg-card border border-card-border rounded-xl p-4 shadow-sm mb-4 animate-pulse">
-            {/* Header */}
-            <div className="flex items-center gap-2 mb-3">
-                <div className="w-4 h-4 rounded-full bg-dim" />
-                <div className="h-3 bg-dim rounded-full w-28" />
-                <div className="h-3 bg-dim rounded-full w-3" />
-                <div className="h-3 bg-dim rounded-full w-36" />
-                <div className="h-3 bg-dim rounded-full w-3" />
-                <div className="h-3 bg-dim rounded-full w-20" />
-            </div>
-            {/* Content */}
-            <div className="space-y-2 mb-3">
-                <div className="h-3.5 bg-dim rounded-full w-full" />
-                <div className="h-3.5 bg-dim rounded-full w-5/6" />
-                <div className="h-3.5 bg-dim rounded-full w-2/3" />
-            </div>
-            {/* Actions */}
-            <div className="flex items-center gap-4 border-t border-card-border/50 pt-3">
-                <div className="flex items-center gap-1.5">
-                    <div className="w-[18px] h-[18px] rounded bg-dim" />
-                    <div className="h-3 bg-dim rounded-full w-6" />
-                </div>
-                <div className="flex items-center gap-1.5">
-                    <div className="w-[18px] h-[18px] rounded bg-dim" />
-                    <div className="h-3 bg-dim rounded-full w-20" />
-                </div>
-            </div>
-        </div>
-    );
-}
 
 interface Community {
     id: string;
