+++ app/(dashboard)/search/page.tsx
@@ -13,6 +13,7 @@ import { BookModal } from "@/components/books";
 import { Toast } from "@/components/ui/Toast";
 import { UserAvatar } from "@/components/ui/UserAvatar";
 import { Users, BookOpen, User, Building2, Search, Loader2 } from "lucide-react";
+import { BookCardSkeleton, UserCardSkeleton, CommunityCardSkeleton } from "@/components/ui/skeletons";
 
 // ─── Types ───────────────────────────────────────────────────────────────────
 
@@ -58,63 +59,7 @@ const TABS: { key: SearchTab; label: string; icon: React.ReactNode }[] = [
     { key: "communities", label: "Comunidades", icon: <Building2 size={16} /> },
 ];
 
-// ─── Skeleton Components ─────────────────────────────────────────────────────
 
-function BookCardSkeleton() {
-    return (
-        <div className="bg-card rounded-2xl border-2 border-card-border p-4 animate-pulse">
-            <div className="aspect-[2/3] bg-dim rounded-xl mb-3" />
-            <div className="space-y-2">
-                <div className="h-4 bg-dim rounded-full w-3/4" />
-                <div className="h-3 bg-dim rounded-full w-1/2" />
-                <div className="flex gap-1 mt-2">
-                    <div className="h-5 bg-dim rounded-full w-16" />
-                    <div className="h-5 bg-dim rounded-full w-12" />
-                </div>
-            </div>
-        </div>
-    );
-}
-
-function UserCardSkeleton() {
-    return (
-        <div className="bg-card rounded-2xl border-2 border-card-border p-6 animate-pulse">
-            <div className="flex items-center gap-4 mb-4">
-                <div className="w-16 h-16 rounded-full bg-dim" />
-                <div className="flex-1 space-y-2">
-                    <div className="h-4 bg-dim rounded-full w-32" />
-                    <div className="h-3 bg-dim rounded-full w-24" />
-                    <div className="h-3 bg-dim rounded-full w-28" />
-                </div>
-            </div>
-            <div className="flex gap-1">
-                <div className="h-5 bg-dim rounded-full w-16" />
-                <div className="h-5 bg-dim rounded-full w-20" />
-                <div className="h-5 bg-dim rounded-full w-14" />
-            </div>
-        </div>
-    );
-}
-
-function CommunityCardSkeleton() {
-    return (
-        <div className="bg-card rounded-2xl border-2 border-card-border p-5 animate-pulse">
-            <div className="flex items-start gap-4 mb-3">
-                <div className="w-16 h-16 rounded-xl bg-dim flex-shrink-0" />
-                <div className="flex-1 space-y-2">
-                    <div className="h-5 bg-dim rounded-full w-3/4" />
-                    <div className="h-3 bg-dim rounded-full w-24" />
-                    <div className="h-4 bg-dim rounded-full w-16" />
-                </div>
-            </div>
-            <div className="space-y-1.5 mb-4">
-                <div className="h-3 bg-dim rounded-full w-full" />
-                <div className="h-3 bg-dim rounded-full w-2/3" />
-            </div>
-            <div className="h-9 bg-dim rounded-lg" />
-        </div>
-    );
-}
 
 // ─── Book Search Card (grid version) ─────────────────────────────────────────
