+++ components/ui/skeletons/index.ts
@@ -1,0 +1,7 @@
+export { BookCardSkeleton } from "./BookCardSkeleton";
+export { PostCardSkeleton } from "./PostCardSkeleton";
+export { UserCardSkeleton } from "./UserCardSkeleton";
+export { CommunityCardSkeleton } from "./CommunityCardSkeleton";
+export { RecommendationSkeleton } from "./RecommendationSkeleton";
+export { BookDetailSkeleton } from "./BookDetailSkeleton";
+export { ChatSkeleton } from "./ChatSkeleton";
